fix example ports; update 3box
