+++ examples/3box/.mm-plugin.json
@@ -1,4 +1,3 @@
 {
-  "root": ".",
   "port": 8083
 }
