+++ examples/3box/README.md
@@ -1,0 +1,4 @@
+
+# 3box Plugin
+
+**NOTE:** This plugin is currently not supported and will be updated in the future.
