+++ examples/bls-signer/index.html
@@ -8,7 +8,7 @@
 
 <body>
   <h1>BLS Signature Plugin Example!</h1>
-  <p>Make sure the plugin bundle is currently hosted on localhost:8083 for this to work.</p>
+  <p>Make sure the plugin bundle is currently hosted on localhost:8084 for this to work.</p>
   <button class="connect">Connect</button>
   <button class="account">Get Public Key</button>
 
@@ -20,7 +20,7 @@
 
 <script>
 
-const origin = 'http://localhost:8083/package.json'
+const origin = 'http://localhost:8084/package.json'
 const pluginOrigin = `wallet_plugin_${origin}`
 
 const connectButton = document.querySelector('button.connect')
