+++ examples/custom-token/package.json
