+++ examples/ping-pong/.mm-plugin.json
@@ -1,4 +1,3 @@
 {
-  "root": ".",
-  "port": 8083
+  "port": 8082
 }
